Refactored cache attribute into a base class and made DittoCacheAttribute sealed
