+++ src/Our.Umbraco.Ditto/ComponentModel/Attributes/DittoCacheAttribute.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Web.Caching;
-using Umbraco.Core;
 
 namespace Our.Umbraco.Ditto
 {
@@ -8,75 +6,6 @@ namespace Our.Umbraco.Ditto
     /// Represents cache configuration data in order to tell Ditto how to cache this property/class
     /// </summary>
     [AttributeUsage(AttributeTargets.Property | AttributeTargets.Class, AllowMultiple = false)]
-    public class DittoCacheAttribute : Attribute
-    {
-        /// <summary>
-        /// Gets or sets the type of the cache key builder.
-        /// </summary>
-        /// <value>
-        /// The type of the cache key builder.
-        /// </value>
-        public Type CacheKeyBuilderType { get; set; }
-
-        /// <summary>
-        /// Gets or sets the properties to cache by.
-        /// </summary>
-        /// <value>
-        /// The cache by property flags.
-        /// </value>
-        public DittoCacheBy CacheBy { get; set; }
-
-        /// <summary>
-        /// Gets or sets the duration of the cache.
-        /// </summary>
-        /// <value>
-        /// The duration of the cache.
-        /// </value>
-        public int CacheDuration { get; set; }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="DittoCacheAttribute"/> class.
-        /// </summary>
-        public DittoCacheAttribute()
-        {
-            CacheBy = Ditto.DefaultCacheBy;
-            CacheDuration = 0;
-        }
-
-        /// <summary>
-        /// Gets the cache item.
-        /// </summary>
-        /// <typeparam name="TOuputType">The type of the ouput type.</typeparam>
-        /// <param name="cacheContext">The cache context.</param>
-        /// <param name="reresher">The reresher.</param>
-        /// <returns></returns>
-        /// <exception cref="System.ApplicationException">Expected a cache key builder of type  + typeof(DittoProcessorCacheKeyBuilder) +  but got  + CacheKeyBuilderType</exception>
-        internal TOuputType GetCacheItem<TOuputType>(DittoCacheContext cacheContext, Func<TOuputType> reresher) 
-        {
-            // If no cache duration set, just run the refresher
-            if (CacheDuration == 0) return reresher();
-
-            // Get the cache key builder type
-            var cacheKeyBuilderType = CacheKeyBuilderType ?? typeof(DittoDefaultCacheKeyBuilder);
-
-            // Check the cache key builder type
-            if (!typeof(DittoCacheKeyBuilder).IsAssignableFrom(cacheKeyBuilderType))
-            {
-                throw new ApplicationException("Expected a cache key builder of type " + typeof(DittoCacheKeyBuilder) + " but got " + CacheKeyBuilderType);
-            }
-
-            // Store a reference to self in the context
-            cacheContext.Attribute = this;
-
-            // Construct the cache key builder
-            var builder = (DittoCacheKeyBuilder)cacheKeyBuilderType.GetInstance();
-            var cacheKey = builder.BuildCacheKey(cacheContext);
-
-            // Get and cache the result
-            return (TOuputType)ApplicationContext.Current.ApplicationCache.RuntimeCache.GetCacheItem(cacheKey,
-                () => reresher(),
-                priority: CacheItemPriority.NotRemovable,
-                timeout: new TimeSpan(0, 0, 0, CacheDuration));
-        }
-    }
+    public sealed class DittoCacheAttribute : DittoCacheableAttribute
+    { }
 }
