+++ src/Our.Umbraco.Ditto/ComponentModel/Attributes/DittoProcessorAttribute.cs
@@ -11,7 +11,7 @@ namespace Our.Umbraco.Ditto
     /// </summary>
     [AttributeUsage(Ditto.ProcessorAttributeTargets)]
     [DittoProcessorMetaData(ValueType = typeof(object), ContextType = typeof(DittoProcessorContext))]
-    public abstract class DittoProcessorAttribute : DittoCacheAttribute
+    public abstract class DittoProcessorAttribute : DittoCacheableAttribute
     {
         /// <summary>
         /// Gets or sets the context.
@@ -106,7 +106,7 @@ namespace Our.Umbraco.Ditto
             Value = value;
             Context = context;
 
-            var ctx = new DittoCacheContext (context.Content, context.TargetType, context.PropertyDescriptor, context.Culture);
+            var ctx = new DittoCacheContext (this, context.Content, context.TargetType, context.PropertyDescriptor, context.Culture);
             return this.GetCacheItem(ctx, this.ProcessValue);
         }
